pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "charon"
version = "0.1.0"
description = "Open-source translational PK platform: SMILES to FIH dose recommendation"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "pydantic>=2.0,<3.0",
    "rdkit-pypi",
    "scipy>=1.11",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "xgboost>=2.0",
    "scikit-learn>=1.3",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/charon"]

[tool.pytest.ini_options]
testpaths = ["tests"]
